refactor(storage): move `query_resumable_upload()`
